// src/api/interface/ICloverApiIncludedElementApp.ts
import type { ICloverApiDeveloper } from './ICloverApiDeveloper';
import type { ICloverApiAggregateRating } from './ICloverApiAggregateRating';
import type { ICloverApiPermission } from './ICloverApiPermission';

export interface ICloverApiIncludedElementApp {
    id: string;
    name: string;
    published: boolean;
    developer: ICloverApiDeveloper;
    description: string;
    tagline: string;
    defaultResponseType: string;
    androidVersion: Record<any, unknown>;
    packageName: string;
    systemApp: boolean;
    hidden: boolean;
    distribution: string;
    filenameIcon: string;
    filenameIconSmall: string;
    filenameIconLarge: string;
    installCount: number;
    sortOrder: number;
    permissions: ICloverApiPermission;
    permissionMerchantRead: boolean;
    permissionMerchantWrite: boolean;
    permissionCustomersRead: boolean;
    permissionCustomersWrite: boolean;
    permissionInventoryRead: boolean;
    permissionInventoryWrite: boolean;
    permissionOrdersRead: boolean;
    permissionOrdersWrite: boolean;
    permissionPaymentsRead: boolean;
    permissionPaymentsWrite: boolean;
    permissionEmployeesRead: boolean;
    permissionEmployeesWrite: boolean;
    permissionProcessCards: boolean;
    permissionMidRead: boolean;
    permissionCustomersAddressRead: boolean;
    permissionCustomersAddressWrite: boolean;
    permissionCustomersBirthdateRead: boolean;
    permissionCustomersBirthdateWrite: boolean;
    permissionCustomersBusinessnameRead: boolean;
    permissionCustomersBusinessnameWrite: boolean;
    permissionCustomersCardRead: boolean;
    permissionCustomersCardWrite: boolean;
    permissionCustomersEmailRead: boolean;
    permissionCustomersEmailWrite: boolean;
    permissionCustomersNoteRead: boolean;
    permissionCustomersNoteWrite: boolean;
    permissionCustomersPhoneRead: boolean;
    permissionCustomersPhoneWrite: boolean;
    permissionCustomersMarketingRead: boolean;
    permissionCustomersMarketingWrite: boolean;
    privacyPolicy: string;
    eula: string;
    approvalStatus: string;
    paidAppHasTrial: boolean;
    approvalStatusModifiedTime: number;
    billingStartTime: number;
    nonCloverBilling: boolean;
    firstPublishedTime: number;
    firstApprovalTime: number;
    firstSubmittedTime: number;
    createdTime: number;
    editorPick: boolean;
    popularity: number;
    allowUninstall: boolean;
    charge: boolean;
    aggregateRating: ICloverApiAggregateRating;
    trialDays: number;
}